server/chatbot/requirements.txt
Flask==2.1.2
requests==2.27.1
python-dotenv==0.20.0
flask-cors==3.0.10
gunicorn==20.1.0
google-generativeai==0.3.2
assemblyai==0.19.1
pydub==0.25.1
SpeechRecognition==3.10.0
